Implement Relation#with_annotation method
Implements `ActiveRecord::Relation#with_annotation`, which accepts a
comment string that will be appeneded to any queries generated by the
relation.

Some examples:

    relation = Post.where(id: 123).with_annotation("metadata string")
    relation.first
    # SELECT "posts".* FROM "posts" WHERE "posts"."id" = 123
    # LIMIT 1 /* metadata string */

    class Tag < ActiveRecord::Base
      scope :foo_annotated, -> { with_annotation("foo") }
    end
    Tag.foo_annotated.with_annotation("-bar").first
    # SELECT "tags".* FROM "tags" LIMIT 1 /* foo-bar */

Also wires up the plumbing so this works with `#update_all` and
`#delete_all` as well.

This feature is useful for instrumentation and general analysis of
queries generated at runtime.